The EWKT parser was dropping leading 0s
When encountering a number and then a decimal point (`.`), the
next match was for an integer. Of course, "0012" is just "12".

When added back as the mantissa of a floating point value, any
leading zeroes would therefore disappear.

This changes the next match after a decimal point to be a collection
of numeric characters, which preserves the zeros. This is then
turned into a number in `to_number`.