Radically drop the resources required to compile the WKT parser
It was building the same combinators over and over again, resulting
in memory usage of ~1.7GB to build and taking ~27s to complete the
build of just the WKT parser on an i7 running at 4GHz.

Creating combinators with `defcombinator` and `defparsec` allows
the combinators used repatedly, such as the coordinate combinators,
to be compiled *once* and used repeatedly at runtime.

Compiling now takes ~175MB and under 2s on the same hardware.